pyproject.toml
[project]
name = "pdf-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "faiss-cpu>=1.13.2",
    "flashrank>=0.2.10",
    "langchain>=1.2.7",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "llama-parse>=0.6.91",
    "openai>=2.15.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.1",
]
